@@ -29,9 +29,9 @@ packages = find:
 install_requires =
     matplotlib
     napari
-    numpy
+    numpy>=1.23
     tinycss2
-python_requires = >=3.9
+python_requires = >=3.10
 include_package_data = True
 package_dir =
     =src
